pyproject.toml
[tool.poetry]
name = "emkademy-vladimir"
version = "0.1.0"
description = ""
authors = ["Vladimir Torres Rivas <vladimir.torres.rivas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
hydra-core = "~=1.3.2"
pydantic = "~=1.10.7"
pandas = "~=2.0"
nltk = "~3.8"
symspellpy = "~=6.7"
fsspec = {version = "~=2025.12", extras = ["gcs"]}
gcsfs = "~=2025.12"
google-cloud-secret-manager = "~=2.16"
fastparquet = "~=2023.4"
dvc = {version = "~=3.64", extras = ["gdrive", "gs"]}
pygit2 = "1.14.1"
dask = {version = "~=2023.5", extras = ["complete"]}
distributed = "~=2023.5"
dask-cloudprovider = {version = "2022.10", extras = ["gcp"]}
dask-ml = "~=2023.3"

[tool.poetry.group.dev.dependencies]
pytest = "~=7.3"
black = "~=23.3"
isort = "~=5.12"
flake8 = "~=6.0"
mypy = "~=1.2"
jupyterlab = "~=4.2"

types-PyYAML = "~=6.0.12.20250915"
types-cachetools ="~=6.2.0.20251022"
types-decorator = "~=5.2.0.20251101"
types-pycurl = "~=7.45.7.20251101"
types-python-dateutil = "~=2.9.0.20251115"
types-requests = "~=2.32.4.20250913"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["cybulde"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.10"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0
